The cursor now accelerates, according to the factors in the cursorcomponent update member. Design question: how fast and after how long should the cursor go fast?

+++ include/shared.hpp
@@ -77,9 +77,8 @@ extern unsigned char id;
 extern std::vector<int> extract_int_string(string);
 
 
-extern int geometricslide(int distance, int geo_factor = 2);
+extern int geometricslide(int distance, float geo_factor = 2);
 extern bool is_pressed(const Uint8 *state_array, std::vector<SDL_Scancode> to_find);
-
 
 template <typename T> int sgn(T val) {
     return (T(0) < val) - (val < T(0));
